Tweak anchor for 27 Languages article
I want the "Broadening our horizons" anchor to be a good lead in to
the 27 Languages article independently of the initial exhortations
to the wider Python community about using pride instead of fear
as a motivator, and focusing on enabling the learning of others
over propagating our own personal preferences.

Moving this paragraph into the preceding section should help with
that, since it's the bridge from the first section to the second,
rather than being an intrinsic piece of the rest of the article.